exhuastiveness: update documentation for th module
