+++ compiler/hash-typecheck/src/exhaustiveness/matrix.rs
@@ -1,5 +1,5 @@
 use crate::{
-    exhaustiveness::structures::PatCtx,
+    exhaustiveness::PatCtx,
     ops::AccessToOps,
     storage::{
         primitives::{ConstructorId, DeconstructedPatId},
@@ -9,9 +9,17 @@ use crate::{
 
 use super::{stack::PatStack, AccessToUsefulnessOps};
 
-/// A 2D matrix.
+/// A 2D matrix which is used to represent the
+/// pattern matrix in order to produce [super::usefulness::Witness]
+/// entries when traversing each row.
+///
+/// Each row can be thought of as a match arm in
+/// the match block, however most rows in the
+/// [Matrix] are generated from when patterns are
+/// specialised or expanded.
 #[derive(Clone)]
 pub struct Matrix {
+    /// The inner rows of the [Matrix].
     pub patterns: Vec<PatStack>,
 }
 
@@ -41,34 +49,34 @@ impl Matrix {
 // /// | false | [_]               |
 // /// | _     | [_, _, ...tail]   |
 // /// ```
-// // impl<'p> fmt::Debug for Matrix<'p> {
-// //     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-// //         writeln!(f)?;
+// impl<'p> fmt::Debug for Matrix<'p> {
+//     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+//         writeln!(f)?;
 
-// //         let Matrix { patterns: m, .. } = self;
-// //         let pretty_printed_matrix: Vec<Vec<String>> =
-// //             m.iter().map(|row| row.iter().map(|pat| format!("{:?}",
-// pat)).collect()).collect();
+//         let Matrix { patterns: m, .. } = self;
+//         let pretty_printed_matrix: Vec<Vec<String>> =
+//             m.iter().map(|row| row.iter().map(|pat| format!("{:?}", pat))
+// .collect()).collect();
 
-// //         let column_count = m.iter().map(|row|
-// row.len()).next().unwrap_or(0); //         assert!(m.iter().all(|row|
-// row.len() == column_count)); //         let column_widths: Vec<usize> =
-// (0..column_count) //             .map(|col|
-// pretty_printed_matrix.iter().map(|row| row[col].len()).max().unwrap_or(0)) //
+//         let column_count = m.iter().map(|row| row.len()).next().unwrap_or(0);
+//         assert!(m.iter().all(|row| row.len() == column_count));
+//         let column_widths: Vec<usize> = (0..column_count)
+//                  .map(|col| pretty_printed_matrix.iter()
+//  .map(|row| row[col].len()).max().unwrap_or(0))
 // .collect();
 
-// //         for row in pretty_printed_matrix {
-// //             write!(f, "|")?;
-// //             for (column, pat_str) in row.into_iter().enumerate() {
-// //                 write!(f, " ")?;
-// //                 write!(f, "{:1$}", pat_str, column_widths[column])?;
-// //                 write!(f, " +")?;
-// //             }
-// //             writeln!(f)?;
-// //         }
-// //         Ok(())
-// //     }
-// // }
+//         for row in pretty_printed_matrix {
+//             write!(f, "|")?;
+//             for (column, pat_str) in row.into_iter().enumerate() {
+//                 write!(f, " ")?;
+//                 write!(f, "{:1$}", pat_str, column_widths[column])?;
+//                 write!(f, " +")?;
+//             }
+//             writeln!(f)?;
+//         }
+//         Ok(())
+//     }
+// }
 
 /// Contains functions related to operations on [Matrix]
 pub struct MatrixOps<'gs, 'ls, 'cd, 's> {
@@ -99,30 +107,31 @@ impl<'gs, 'ls, 'cd, 's> MatrixOps<'gs, 'ls, 'cd, 's> {
         }
     }
 
-    /// This computes `S(constructor, self)`. See top of the file for
-    /// explanations.
+    /// This computes `S(constructor, matrix)`.
     pub(crate) fn specialise_ctor(
         &self,
         ctx: PatCtx,
         matrix: &Matrix,
         ctor_id: ConstructorId,
     ) -> Matrix {
-        let reader = self.reader();
-        let mut s_matrix = Matrix::empty();
-
+        let mut specialised_matrix = Matrix::empty();
         let ctor = self.constructor_store().get(ctor_id);
+        let reader = self.reader();
 
+        // Iterate on each row, and specialise the `head` of
+        // each row within the matrix with the provided constructor,
+        // the results of the specialisation as new rows in
+        // the matrix.
         for row in &matrix.patterns {
             let head = reader.get_deconstructed_pat(row.head());
-
             let other = self.constructor_store().get(head.ctor());
 
             if self.constructor_ops().is_covered_by(ctx, &ctor, &other) {
                 let new_row = self.stack_ops().pop_head_constructor(ctx, row, ctor_id);
-                self.push(&mut s_matrix, new_row);
+                self.push(&mut specialised_matrix, new_row);
             }
         }
 
-        s_matrix
+        specialised_matrix
     }
 }
